Remove IPA API version, SSL verify to True, cleanup
In IPA 4.0+ the API version isn't mandatory so there is no need
to try to fake the server out.

Since the IPA CA certificate is added to the global CA store
we can set verify to True and have SSL validation done
automatically.

Add the command to the "Ignored" errors so we can know what
failed.

Remove unused debug statement.