Remove unused files, reorder nova-setup.sh, use openstack cli
Install python-novajoin via the package list rather than in the
setup script.

Reorder the setup script to make it easier to re-use it to spin
up another instance (by removing lines).

Remove some more unnecessary lines, like configuring
cloud-config.json

Use the openstack cli command in more places and make use of the
-f value -c <column> to get rid of a lot of the pipes and awks

When I moved the package install into the list the ownership of
/var/log/novajoin was root:root when the services needed to start
which of course failed. This is a stop-gap while I investigate
that.